content-gen/src/requirements.txt
# Content Generation Solution Accelerator - Python Dependencies

# Web Framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.17.0

# Microsoft Agent Framework
agent-framework>=0.1.0

# Azure SDKs
azure-identity>=1.17.0
azure-cosmos>=4.7.0
azure-storage-blob>=12.22.0
azure-ai-contentsafety>=1.0.0

# OpenAI
openai>=1.45.0

# Data Validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
